Move script to dev/ [skip ci]
